LayoutTests/inspector/console/console-table-expected.txt

CONSOLE MESSAGE: line 3: 
CONSOLE MESSAGE: line 3: apple,orange,banana
CONSOLE MESSAGE: line 3: [object Object]
CONSOLE MESSAGE: line 3: John,Smith,Jane,Doe,Emily,Jones
CONSOLE MESSAGE: line 3: [object Object],[object Object],[object Object]
CONSOLE MESSAGE: line 3: [object Object],[object Object],[object Object]

STEP: console.table([])
{
  "_listeners": null,
  "_source": "console-api",
  "_level": "log",
  "_messageText": "",
  "_type": "table",
  "_url": null,
  "_line": 3,
  "_column": 14,
  "_repeatCount": 1,
  "_parameters": [
    {
      "_type": "object",
      "_subtype": "array",
      "_objectId": "<filtered>",
      "_description": "Array",
      "_hasChildren": true,
      "_size": 0,
      "_preview": {
        "_listeners": null,
        "_type": "object",
        "_subtype": "array",
        "_description": "Array",
        "_lossless": true,
        "_overflow": false,
        "_size": 0,
        "_properties": [],
        "_entries": null
      }
    }
  ],
  "_stackTrace": "<filtered>",
  "_request": null
}

STEP: console.table(['apple', 'orange', 'banana'])
{
  "_listeners": null,
  "_source": "console-api",
  "_level": "log",
  "_messageText": "apple,orange,banana",
  "_type": "table",
  "_url": null,
  "_line": 3,
  "_column": 14,
  "_repeatCount": 1,
  "_parameters": [
    {
      "_type": "object",
      "_subtype": "array",
      "_objectId": "<filtered>",
      "_description": "Array",
      "_hasChildren": true,
      "_size": 3,
      "_preview": {
        "_listeners": null,
        "_type": "object",
        "_subtype": "array",
        "_description": "Array",
        "_lossless": true,
        "_overflow": false,
        "_size": 3,
        "_properties": [
          {
            "_listeners": null,
            "_name": "0",
            "_type": "string",
            "_value": "apple"
          },
          {
            "_listeners": null,
            "_name": "1",
            "_type": "string",
            "_value": "orange"
          },
          {
            "_listeners": null,
            "_name": "2",
            "_type": "string",
            "_value": "banana"
          }
        ],
        "_entries": null
      }
    }
  ],
  "_stackTrace": "<filtered>",
  "_request": null
}

STEP: console.table({firstName: 'John', lastName: 'Smith'})
{
  "_listeners": null,
  "_source": "console-api",
  "_level": "log",
  "_messageText": "[object Object]",
  "_type": "table",
  "_url": null,
  "_line": 3,
  "_column": 14,
  "_repeatCount": 1,
  "_parameters": [
    {
      "_type": "object",
      "_objectId": "<filtered>",
      "_description": "Object",
      "_hasChildren": true,
      "_preview": {
        "_listeners": null,
        "_type": "object",
        "_description": "Object",
        "_lossless": true,
        "_overflow": false,
        "_properties": [
          {
            "_listeners": null,
            "_name": "firstName",
            "_type": "string",
            "_value": "John"
          },
          {
            "_listeners": null,
            "_name": "lastName",
            "_type": "string",
            "_value": "Smith"
          }
        ],
        "_entries": null
      }
    }
  ],
  "_stackTrace": "<filtered>",
  "_request": null
}

STEP: console.table([['John', 'Smith'], ['Jane', 'Doe'], ['Emily', 'Jones']])
{
  "_listeners": null,
  "_source": "console-api",
  "_level": "log",
  "_messageText": "John,Smith,Jane,Doe,Emily,Jones",
  "_type": "table",
  "_url": null,
  "_line": 3,
  "_column": 14,
  "_repeatCount": 1,
  "_parameters": [
    {
      "_type": "object",
      "_subtype": "array",
      "_objectId": "<filtered>",
      "_description": "Array",
      "_hasChildren": true,
      "_size": 3,
      "_preview": {
        "_listeners": null,
        "_type": "object",
        "_subtype": "array",
        "_description": "Array",
        "_lossless": true,
        "_overflow": false,
        "_size": 3,
        "_properties": [
          {
            "_listeners": null,
            "_name": "0",
            "_type": "object",
            "_subtype": "array",
            "_valuePreview": {
              "_listeners": null,
              "_type": "object",
              "_subtype": "array",
              "_description": "Array",
              "_lossless": true,
              "_overflow": false,
              "_size": 2,
              "_properties": [
                {
                  "_listeners": null,
                  "_name": "0",
                  "_type": "string",
                  "_value": "John"
                },
                {
                  "_listeners": null,
                  "_name": "1",
                  "_type": "string",
                  "_value": "Smith"
                }
              ],
              "_entries": null
            }
          },
          {
            "_listeners": null,
            "_name": "1",
            "_type": "object",
            "_subtype": "array",
            "_valuePreview": {
              "_listeners": null,
              "_type": "object",
              "_subtype": "array",
              "_description": "Array",
              "_lossless": true,
              "_overflow": false,
              "_size": 2,
              "_properties": [
                {
                  "_listeners": null,
                  "_name": "0",
                  "_type": "string",
                  "_value": "Jane"
                },
                {
                  "_listeners": null,
                  "_name": "1",
                  "_type": "string",
                  "_value": "Doe"
                }
              ],
              "_entries": null
            }
          },
          {
            "_listeners": null,
            "_name": "2",
            "_type": "object",
            "_subtype": "array",
            "_valuePreview": {
              "_listeners": null,
              "_type": "object",
              "_subtype": "array",
              "_description": "Array",
              "_lossless": true,
              "_overflow": false,
              "_size": 2,
              "_properties": [
                {
                  "_listeners": null,
                  "_name": "0",
                  "_type": "string",
                  "_value": "Emily"
                },
                {
                  "_listeners": null,
                  "_name": "1",
                  "_type": "string",
                  "_value": "Jones"
                }
              ],
              "_entries": null
            }
          }
        ],
        "_entries": null
      }
    }
  ],
  "_stackTrace": "<filtered>",
  "_request": null
}

STEP: console.table([john, jane, emily])
{
  "_listeners": null,
  "_source": "console-api",
  "_level": "log",
  "_messageText": "[object Object],[object Object],[object Object]",
  "_type": "table",
  "_url": null,
  "_line": 3,
  "_column": 14,
  "_repeatCount": 1,
  "_parameters": [
    {
      "_type": "object",
      "_subtype": "array",
      "_objectId": "<filtered>",
      "_description": "Array",
      "_hasChildren": true,
      "_size": 3,
      "_preview": {
        "_listeners": null,
        "_type": "object",
        "_subtype": "array",
        "_description": "Array",
        "_lossless": false,
        "_overflow": false,
        "_size": 3,
        "_properties": [
          {
            "_listeners": null,
            "_name": "0",
            "_type": "object",
            "_valuePreview": {
              "_listeners": null,
              "_type": "object",
              "_description": "Person",
              "_lossless": false,
              "_overflow": false,
              "_properties": [
                {
                  "_listeners": null,
                  "_name": "firstName",
                  "_type": "string",
                  "_value": "John"
                },
                {
                  "_listeners": null,
                  "_name": "lastName",
                  "_type": "string",
                  "_value": "Smith"
                }
              ],
              "_entries": null
            }
          },
          {
            "_listeners": null,
            "_name": "1",
            "_type": "object",
            "_valuePreview": {
              "_listeners": null,
              "_type": "object",
              "_description": "Person",
              "_lossless": false,
              "_overflow": false,
              "_properties": [
                {
                  "_listeners": null,
                  "_name": "firstName",
                  "_type": "string",
                  "_value": "Jane"
                },
                {
                  "_listeners": null,
                  "_name": "lastName",
                  "_type": "string",
                  "_value": "Doe"
                }
              ],
              "_entries": null
            }
          },
          {
            "_listeners": null,
            "_name": "2",
            "_type": "object",
            "_valuePreview": {
              "_listeners": null,
              "_type": "object",
              "_description": "Person",
              "_lossless": false,
              "_overflow": false,
              "_properties": [
                {
                  "_listeners": null,
                  "_name": "firstName",
                  "_type": "string",
                  "_value": "Emily"
                },
                {
                  "_listeners": null,
                  "_name": "lastName",
                  "_type": "string",
                  "_value": "Jones"
                }
              ],
              "_entries": null
            }
          }
        ],
        "_entries": null
      }
    }
  ],
  "_stackTrace": "<filtered>",
  "_request": null
}

STEP: console.table([john, jane, emily], ['firstName'])
{
  "_listeners": null,
  "_source": "console-api",
  "_level": "log",
  "_messageText": "[object Object],[object Object],[object Object]",
  "_type": "table",
  "_url": null,
  "_line": 3,
  "_column": 14,
  "_repeatCount": 1,
  "_parameters": [
    {
      "_type": "object",
      "_subtype": "array",
      "_objectId": "<filtered>",
      "_description": "Array",
      "_hasChildren": true,
      "_size": 3,
      "_preview": {
        "_listeners": null,
        "_type": "object",
        "_subtype": "array",
        "_description": "Array",
        "_lossless": false,
        "_overflow": false,
        "_size": 3,
        "_properties": [
          {
            "_listeners": null,
            "_name": "0",
            "_type": "object",
            "_valuePreview": {
              "_listeners": null,
              "_type": "object",
              "_description": "Person",
              "_lossless": false,
              "_overflow": false,
              "_properties": [
                {
                  "_listeners": null,
                  "_name": "firstName",
                  "_type": "string",
                  "_value": "John"
                },
                {
                  "_listeners": null,
                  "_name": "lastName",
                  "_type": "string",
                  "_value": "Smith"
                }
              ],
              "_entries": null
            }
          },
          {
            "_listeners": null,
            "_name": "1",
            "_type": "object",
            "_valuePreview": {
              "_listeners": null,
              "_type": "object",
              "_description": "Person",
              "_lossless": false,
              "_overflow": false,
              "_properties": [
                {
                  "_listeners": null,
                  "_name": "firstName",
                  "_type": "string",
                  "_value": "Jane"
                },
                {
                  "_listeners": null,
                  "_name": "lastName",
                  "_type": "string",
                  "_value": "Doe"
                }
              ],
              "_entries": null
            }
          },
          {
            "_listeners": null,
            "_name": "2",
            "_type": "object",
            "_valuePreview": {
              "_listeners": null,
              "_type": "object",
              "_description": "Person",
              "_lossless": false,
              "_overflow": false,
              "_properties": [
                {
                  "_listeners": null,
                  "_name": "firstName",
                  "_type": "string",
                  "_value": "Emily"
                },
                {
                  "_listeners": null,
                  "_name": "lastName",
                  "_type": "string",
                  "_value": "Jones"
                }
              ],
              "_entries": null
            }
          }
        ],
        "_entries": null
      }
    },
    {
      "_type": "object",
      "_subtype": "array",
      "_objectId": "<filtered>",
      "_description": "Array",
      "_hasChildren": true,
      "_size": 1,
      "_preview": {
        "_listeners": null,
        "_type": "object",
        "_subtype": "array",
        "_description": "Array",
        "_lossless": true,
        "_overflow": false,
        "_size": 1,
        "_properties": [
          {
            "_listeners": null,
            "_name": "0",
            "_type": "string",
            "_value": "firstName"
          }
        ],
        "_entries": null
      }
    }
  ],
  "_stackTrace": "<filtered>",
  "_request": null
}
